requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
openai>=1.12.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pytesseract>=0.3.10